requirements.txt
google-genai
python-dotenv
youtube-transcript-api
requests
# feedparser==6.0.10
